fix: update v2 design for claim-mining-reward
This implements a 2 part query process where getUserConfig() is called from the main utils.ts for the user configuration, and getScriptConfig() is called locally in the script for any relevant information.

The motivation is to create an easy way to standardize the address, derivation for child keys, and network settings in one configuration used across all scripts. The selected city returned can be used to query relevant contract info for submitting transactions.

The script config covers any locally used variables, e.g. in the case of claim-mining-reward, the start and end block to claim

This removes the "check then claim" logic in favor of simply submitting the TX for now until the API is updated.